+++ Leomard/Models/Lemmy/Forms/CreateCommentReport.swift
@@ -1,0 +1,13 @@
+//
+//  CreateCommentReport.swift
+//  Leomard
+//
+//  Created by Konrad Figura on 22/07/2023.
+//
+
+import Foundation
+
+struct CreateCommentReport: Codable {
+    let commentId: Int
+    let reason: String
+}