+++ Leomard/Models/Lemmy/Report/CommentReportResponse.swift
@@ -1,0 +1,12 @@
+//
+//  CommentReportResponse.swift
+//  Leomard
+//
+//  Created by Konrad Figura on 22/07/2023.
+//
+
+import Foundation
+
+struct CommentReportResponse: Codable {
+    let commentReportView: CommentReportView
+}