+++ Leomard/Views/Components/ContextMenus/PostContextMenu.swift
@@ -9,6 +9,7 @@ import Foundation
 import SwiftUI
 
 struct PostContextMenu: View {
+    let contentView: ContentView
     let postView: PostView
     @Environment(\.openURL) var openURL
     
@@ -57,14 +58,13 @@ struct PostContextMenu: View {
         Divider()
         ShareLink(item: URL(string: postView.post .apId)!) {
             Text("Share")
-        }
+        }        
         Button(action: {
-            
+            contentView.startReport(postView.post)
         }) {
             Text("Report")
                 .padding()
         }
-        .disabled(true)
     }
     
     func saveImage(imageLink: String) {