+++ Leomard/Views/ContentView.swift
@@ -9,8 +9,6 @@ import SwiftUI
 import MarkdownUI
 
 struct ContentView: View {
-    //@StateObject var userPreferences: UserPreferences
-    
     @State var currentSelection: Option = Option(id: 0, title: "Home", imageName: "house")
     var options: [Option] = [
         .init(id: 0, title: "Feed", imageName: "house"),
@@ -44,6 +42,15 @@ struct ContentView: View {
     @State var addingNewUser: Bool = false
     
     @State var interactionEnabled: Bool = true
+    
+    @State var reportingPost: Bool = false
+    @State var reportedPost: Post? = nil
+    @State var reportingComment: Bool = false
+    @State var reportedComment: Comment? = nil
+    
+    @State var reportReason: String = ""
+    @State var reportSent: Bool = false
+    @State var reportSuccess: Bool = false
     
     var appIconBadge = AppAlertBadge()
     
@@ -115,6 +122,51 @@ struct ContentView: View {
         }
         .allowsHitTesting(interactionEnabled)
         .overlay(Color.gray.opacity(interactionEnabled ? 0 : 0.5))
+        .alert("Report Post", isPresented: $reportingPost, actions: {
+            TextField("Reason", text: $reportReason)
+            Spacer()
+            Button("Report", role: .destructive) {
+                self.postService!.report(post: reportedPost!, reason: reportReason) { result in
+                    switch result {
+                    case .success(_):
+                        reportSuccess = true
+                        reportSent = true
+                    case .failure(let error):
+                        print(error)
+                        reportSuccess = false
+                        reportSent = true
+                    }
+                }
+            }
+            .disabled(reportReason.count == 0)
+            Button("Cancel", role: .cancel) {}
+        }, message: {
+            Text("State the reason of your report:")
+        })
+        .alert("Report Comment", isPresented: $reportingComment, actions: {
+            TextField("Reason", text: $reportReason)
+            Spacer()
+            Button("Report", role: .destructive) {
+                self.commentService!.report(comment: reportedComment!, reason: reportReason) { result in
+                    switch result {
+                    case .success(_):
+                        reportSuccess = true
+                        reportSent = true
+                    case .failure(let error):
+                        print(error)
+                        reportSuccess = false
+                        reportSent = true
+                    }
+                }
+            }
+            .disabled(reportReason.count == 0)
+            Button("Cancel", role: .cancel) {}
+        }, message: {
+            Text("State the reason of your report:")
+        })
+        .alert(reportSuccess ? "Success!" : "Error",
+               isPresented: $reportSent, actions: {},
+               message: { Text(reportSuccess ? "Your report has been sent." : "Failed to send report. Try again later.") })
     }
     
     /// - Returns: A view reflecting whether user is logged in to a profile or user needs to be prompted to log in.
@@ -364,5 +416,15 @@ struct ContentView: View {
     
     func toggleInteraction(_ enabled: Bool) {
         interactionEnabled = enabled
+    }
+    
+    func startReport(_ post: Post) {
+        self.reportedPost = post
+        reportingPost = true
+    }
+    
+    func startReport(_ comment: Comment) {
+        self.reportedComment = comment
+        reportingComment = true
     }
 }