+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - Image uploading
 - Search within community
 - OP indicator in the comments
+- Post reporting
+- Comment reporting
 
 ### Changes